rewriting server
