+++ .gitignore
@@ -1,6 +1,5 @@
 # Dependency directory
 node_modules/
-*-lock*
 
 # Runtime data
 pids
