+++ api/utils/epub_manager.js
