+++ package.json
@@ -1,11 +1,11 @@
 {
   "name": "manga2kindle_api",
-  "version": "0.7.3",
+  "version": "0.8.0",
   "description": "",
-  "main": "app.js",
+  "main": "index.js",
   "scripts": {
-    "start": "node app.js",
-    "test": "npx standard && echo \"Error: no test specified\" && exit 1"
+    "start": "node index.js",
+    "test": "npx standard && npx mocha"
   },
   "author": "Eduardo Fernandez",
   "license": "unlicensed",
@@ -13,15 +13,20 @@
     "adm-zip": "^0.4.14",
     "archiver": "^4.0.1",
     "body-parser": "^1.19.0",
+    "debug": "^4.1.1",
     "dotenv": "^8.0.0",
     "express": "^4.17.1",
-    "express-fileupload": "^1.1.6",
+    "express-fileupload": "^1.1.7-alpha.3",
+    "morgan": "^1.10.0",
     "nodemailer": "^6.1.1",
     "path": "^0.12.7",
-    "pg": "^8.0.2",
+    "pg": "^8.1.0",
     "shelljs": "^0.8.3",
-    "standard": "^14.3.3",
-    "worker-farm": "^1.7.0",
+    "winston": "^3.2.1",
     "xml2js": "^0.4.23"
+  },
+  "devDependencies": {
+    "mocha": "^7.1.2",
+    "standard": "^14.3.3"
   }
 }
