+++ test/test.js
@@ -1,0 +1,8 @@
+/**
+ * No tests for now, check this:
+ *
+ * Test library: https://mochajs.org/ (already installed, the website has a really usefull guide)
+ * A generic test guide: https://medium.com/@me_37286/yoni-goldberg-javascript-nodejs-testing-best-practices-2b98924c9347
+ */
+
+console.log('Error: no test specified')
